Allow the sessions to get and tune the socket buffers
Getting the socket buffer size of very important for the session to know
how big it needs to create its buffers (e.g. when compressing we need to
know how much we can write at once on the socket).
Tuning the socket buffers sizes is useful to decrease the memory usage
of sessions.