Introduce wakeup machanism
It's useful when the session needs a lot of time (>10ms) to process the
request. This way we delegate the processing to a worker thread and
when's done, we wake up the session and write the result.

Be aware that the session might be wakeup also by the kernel, so you
need to make sure that the worker finished the job before you start
writing any data.